Increased readibilty of code

+++ src/main/java/com/strandls/utility/controller/UtilityController.java
@@ -371,20 +371,22 @@ public class UtilityController {
 			return Response.status(Status.BAD_REQUEST).entity(e.getMessage()).build();
 		}
 	}
-	
+
 	@POST
-	@Path(ApiConstants.HOMEPAGE + ApiConstants.MINI_GALLERY +ApiConstants.CREATE )
+	@Path(ApiConstants.HOMEPAGE + ApiConstants.MINI_GALLERY + ApiConstants.CREATE)
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	@ValidateUser
 
 	@ApiOperation(value = "Creates a new mini gallery", notes = "Return created mini gallery", response = GalleryConfig.class)
-	@ApiResponses(value = { @ApiResponse(code = 400, message = "Unable to create mini gallery", response = String.class)})
+	@ApiResponses(value = {
+			@ApiResponse(code = 400, message = "Unable to create mini gallery", response = String.class) })
 
-	public Response createMiniGallery(@Context HttpServletRequest request, @ApiParam(name = "miniGalleryData") GalleryConfig miniGalleryData) {
+	public Response createMiniGallery(@Context HttpServletRequest request,
+			@ApiParam(name = "miniGalleryData") GalleryConfig miniGalleryData) {
 		try {
 			GalleryConfig result = utilityService.createMiniGallery(request, miniGalleryData);
-			if (result!=null)
+			if (result != null)
 				return Response.status(Status.OK).entity(result).build();
 			return Response.status(Status.NOT_FOUND).build();
 
@@ -392,9 +394,9 @@ public class UtilityController {
 			return Response.status(Status.BAD_REQUEST).entity(e.getMessage()).build();
 		}
 	}
-	
+
 	@PUT
-	@Path(ApiConstants.HOMEPAGE + ApiConstants.MINI_GALLERY +ApiConstants.EDIT + "/{galleryId}")
+	@Path(ApiConstants.HOMEPAGE + ApiConstants.MINI_GALLERY + ApiConstants.EDIT + "/{galleryId}")
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 
@@ -407,6 +409,9 @@ public class UtilityController {
 	public Response editMiniGallery(@Context HttpServletRequest request, @PathParam("galleryId") String galleryId,
 			@ApiParam(name = "editData") GalleryConfig editData) {
 		try {
+			if (galleryId == null) {
+				return Response.status(Status.BAD_REQUEST).entity("Gallery Id cannot be null").build();
+			}
 			Long gId = Long.parseLong(galleryId);
 			GalleryConfig result = utilityService.editMiniGallery(request, gId, editData);
 			if (result != null)
@@ -417,7 +422,7 @@ public class UtilityController {
 			return Response.status(Status.BAD_REQUEST).entity(e.getMessage()).build();
 		}
 	}
-	
+
 	@DELETE
 	@Path(ApiConstants.HOMEPAGE + ApiConstants.MINI_GALLERY + ApiConstants.REMOVE + "/{galleryId}")
 	@Consumes(MediaType.TEXT_PLAIN)
@@ -426,15 +431,17 @@ public class UtilityController {
 	@ValidateUser
 
 	@ApiOperation(value = "Delete mini gallery data", notes = "returns null", response = Void.class)
-	@ApiResponses(value = {
-			@ApiResponse(code = 400, message = "unable to delete the data", response = String.class) })
+	@ApiResponses(value = { @ApiResponse(code = 400, message = "unable to delete the data", response = String.class) })
 
-	public Response removeMiniGalleryData(@Context HttpServletRequest request, @PathParam("galleryId") String galleryId) {
+	public Response removeMiniGalleryData(@Context HttpServletRequest request,
+			@PathParam("galleryId") String galleryId) {
 		try {
+			if (galleryId == null) {
+				return Response.status(Status.BAD_REQUEST).entity("Gallery Id cannot be null").build();
+			}
 			Long gId = Long.parseLong(galleryId);
-			System.out.println(gId.toString());
 			Boolean result = utilityService.removeMiniGallery(request, gId);
-			if (result != false)
+			if (Boolean.TRUE.equals(result))
 				return Response.status(Status.OK).build();
 			return Response.status(Status.NOT_FOUND).build();
 
@@ -491,9 +498,9 @@ public class UtilityController {
 			return Response.status(Status.BAD_REQUEST).entity(e.getMessage()).build();
 		}
 	}
-	
+
 	@PUT
-	@Path(ApiConstants.HOMEPAGE + ApiConstants.EDIT + ApiConstants.MINI_SLIDER +"/{galleryId}")
+	@Path(ApiConstants.HOMEPAGE + ApiConstants.EDIT + ApiConstants.MINI_SLIDER + "/{galleryId}")
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 
@@ -507,7 +514,7 @@ public class UtilityController {
 			@ApiParam(name = "editData") Map<Long, List<MiniGallerySlider>> editData) {
 		try {
 			Long gId = Long.parseLong(galleryId);
-			HomePageData result = utilityService.editMiniHomePage(request ,gId, editData);
+			HomePageData result = utilityService.editMiniHomePage(request, gId, editData);
 			if (result != null)
 				return Response.status(Status.OK).entity(result).build();
 			return Response.status(Status.NOT_FOUND).build();
@@ -516,7 +523,7 @@ public class UtilityController {
 			return Response.status(Status.BAD_REQUEST).entity(e.getMessage()).build();
 		}
 	}
-	
+
 	@DELETE
 	@Path(ApiConstants.HOMEPAGE + ApiConstants.REMOVE + ApiConstants.MINI_SLIDER + "/{galleryId}")
 	@Consumes(MediaType.TEXT_PLAIN)
@@ -560,9 +567,9 @@ public class UtilityController {
 			return Response.status(Status.BAD_REQUEST).entity(e.getMessage()).build();
 		}
 	}
-	
+
 	@PUT
-	@Path(ApiConstants.HOMEPAGE + ApiConstants.MINI_SLIDER +ApiConstants.REORDERING)
+	@Path(ApiConstants.HOMEPAGE + ApiConstants.MINI_SLIDER + ApiConstants.REORDERING)
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 
@@ -642,7 +649,7 @@ public class UtilityController {
 	}
 
 	@GET
-	@Path(ApiConstants.RESOURCE )
+	@Path(ApiConstants.RESOURCE)
 	@Consumes(MediaType.TEXT_PLAIN)
 	@Produces(MediaType.APPLICATION_JSON)
